+++ deltachat-ffi/src/lib.rs
@@ -4452,11 +4452,13 @@ pub unsafe extern "C" fn dc_accounts_get_event_emitter(
 mod jsonrpc {
     use super::*;
     use deltachat_jsonrpc::api::CommandApi;
+    use deltachat_jsonrpc::events::event_to_json_rpc_notification;
     use deltachat_jsonrpc::yerpc::{OutReceiver, RpcClient, RpcSession};
 
     pub struct dc_jsonrpc_instance_t {
         receiver: OutReceiver,
         handle: RpcSession<CommandApi>,
+        event_thread: JoinHandle<Result<(), anyhow::Error>>,
     }
 
     #[no_mangle]
@@ -4472,9 +4474,36 @@ mod jsonrpc {
             deltachat_jsonrpc::api::CommandApi::from_arc((*account_manager).inner.clone());
 
         let (request_handle, receiver) = RpcClient::new();
+        let request_handle2 = request_handle.clone();
         let handle = RpcSession::new(request_handle, cmd_api);
 
-        let instance = dc_jsonrpc_instance_t { receiver, handle };
+        let events = block_on({
+            async {
+                let am = (*account_manager).inner.clone();
+                let ev = am.read().await.get_event_emitter();
+                drop(am);
+                ev
+            }
+        });
+
+        let event_thread = spawn({
+            async move {
+                while let Some(event) = events.recv().await {
+                    let event = event_to_json_rpc_notification(event);
+                    request_handle2
+                        .send_notification("event", Some(event))
+                        .await?;
+                }
+                let res: Result<(), anyhow::Error> = Ok(());
+                res
+            }
+        });
+
+        let instance = dc_jsonrpc_instance_t {
+            receiver,
+            handle,
+            event_thread,
+        };
 
         Box::into_raw(Box::new(instance))
     }
@@ -4485,7 +4514,7 @@ mod jsonrpc {
             eprintln!("ignoring careless call to dc_jsonrpc_unref()");
             return;
         }
-
+        (*jsonrpc_instance).event_thread.abort();
         Box::from_raw(jsonrpc_instance);
     }